+++ paddle/framework/tensor_util.cu
@@ -1,0 +1,1 @@
+./tensor_util.cc
